Add a default mock user and remove sessions concept from mock
In this implementation, a default mock user was created and is the one used for
mocking purposes.

Even being possible to signUp/signIn with new users, the set/enumerate/getPref/setPref
functions always assumes that there is a logged-in user, the default mock user.
In fact, the concep of sessions was removed from the mock.