Implement key pair generation as a promise outside the constructor
The generation of public and secret key pair is a time-consuming operation
that should not be implemented inside the constructor of the class.

As an Account object without key pairs has no sense, this implementation
moves the key pair generation to a static method that should be called
before creating a new Account object.