Improve account to support symmetric encryption
From now, the messages (entries) will be encrypted using a
symmetric encryption schema with a random key generated when
the user signs up. The public and secret keypair generated using
the username and password (from now, the keypair) will be used
to encrypt this key, which will be sent to the server and keep
there.

When the user signs in, the key must be requested and decrypted
using the keypair, and the entries must be decrypted using this
key. With this schema, it will be possible for the user to change
te password and still be able to decrypt the entries.